Alibaba nodepool support with Scaling Group CreateCluster
