+++ Gopkg.lock
@@ -147,7 +147,7 @@
   revision = "de5bf2ad457846296e2031421a34e2568e304e35"
 
 [[projects]]
-  digest = "1:379172d114b64135b9942eddaa15d39f682bbdda5fe7c0207856471490ac5213"
+  digest = "1:e0287a471dde40bef7acd33ab1a32de76cc561ec894641c7f7795a33e7f8a61e"
   name = "github.com/aliyun/alibaba-cloud-sdk-go"
   packages = [
     "sdk",
@@ -161,6 +161,7 @@
     "sdk/utils",
     "services/cs",
     "services/ecs",
+    "services/ess",
   ]
   pruneopts = "NUT"
   revision = "1c5a1c93a9c1e3da49ea0aa1f095a204b075c6b8"
@@ -2197,6 +2198,7 @@
     "github.com/aliyun/alibaba-cloud-sdk-go/sdk/requests",
     "github.com/aliyun/alibaba-cloud-sdk-go/services/cs",
     "github.com/aliyun/alibaba-cloud-sdk-go/services/ecs",
+    "github.com/aliyun/alibaba-cloud-sdk-go/services/ess",
     "github.com/aliyun/aliyun-oss-go-sdk/oss",
     "github.com/antihax/optional",
     "github.com/aokoli/goutils",
