+++ pkg/cluster/acsk/action/action_upload_ssh_key.go
@@ -1,0 +1,74 @@
+// Copyright © 2018 Banzai Cloud
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package action
+
+import (
+	"strings"
+
+	"github.com/aliyun/alibaba-cloud-sdk-go/sdk/requests"
+	"github.com/aliyun/alibaba-cloud-sdk-go/services/ecs"
+	"github.com/banzaicloud/pipeline/secret"
+	"github.com/sirupsen/logrus"
+)
+
+// UploadSSHKeyAction describes how to upload an SSH key
+type UploadSSHKeyAction struct {
+	context   *ACSKClusterCreateContext
+	sshSecret *secret.SecretItemResponse
+	log       logrus.FieldLogger
+}
+
+// NewUploadSSHKeyAction creates a new UploadSSHKeyAction
+func NewUploadSSHKeyAction(log logrus.FieldLogger, context *ACSKClusterCreateContext, sshSecret *secret.SecretItemResponse) *UploadSSHKeyAction {
+	return &UploadSSHKeyAction{
+		context:   context,
+		sshSecret: sshSecret,
+		log:       log,
+	}
+}
+
+// GetName returns the name of this UploadSSHKeyAction
+func (a *UploadSSHKeyAction) GetName() string {
+	return "UploadSSHKeyAction"
+}
+
+// ExecuteAction executes this UploadSSHKeyAction
+func (a *UploadSSHKeyAction) ExecuteAction(input interface{}) (interface{}, error) {
+	a.log.Info("EXECUTE UploadSSHKeyAction")
+	ecsClient := a.context.ECSClient
+
+	req := ecs.CreateImportKeyPairRequest()
+	req.SetScheme(requests.HTTPS)
+	req.KeyPairName = a.context.AlibabaClusterCreateParams.Name
+	req.PublicKeyBody = strings.TrimSpace(secret.NewSSHKeyPair(a.sshSecret).PublicKeyData)
+	req.RegionId = a.context.AlibabaClusterCreateParams.RegionID
+
+	return ecsClient.ImportKeyPair(req)
+}
+
+// UndoAction rolls back this UploadSSHKeyAction
+func (a *UploadSSHKeyAction) UndoAction() (err error) {
+	a.log.Info("EXECUTE UNDO UploadSSHKeyAction")
+	//delete uploaded keypair
+	ecsClient := a.context.ECSClient
+
+	req := ecs.CreateDeleteKeyPairsRequest()
+	req.SetScheme(requests.HTTPS)
+	req.KeyPairNames = a.context.AlibabaClusterCreateParams.Name
+	req.RegionId = a.context.AlibabaClusterCreateParams.RegionID
+
+	_, err = ecsClient.DeleteKeyPairs(req)
+	return
+}
